fixed settings validation for refresh, access

+++ app/core/config.py
@@ -13,8 +13,8 @@ class Settings(BaseSettings):
     SECRET_KEY: str
     REFRESH_SECRET_KEY: str
     ALGORITHM: str
-    ACCESS_TOKEN_EXPIRE_MINUTES: int
-    REFRESH_TOKEN_EXPIRE_DAYS: int
+    ACCESS_TOKEN_EXPIRE_MINUTES: float
+    REFRESH_TOKEN_EXPIRE_DAYS: float
     EMAIL_USER: str
     EMAIL_PASSWORD: str
     SMTP_SERVER: str
